Client: Default Unlock fps is now off, and Locked fps now caps to 65 instead of 33

+++ Orion+#/Source/Client/Source/Modules/Systems/C_Crafting.cs
@@ -501,9 +501,9 @@ namespace Engine
 			
 			for (var i = 1; i <= Constants.MAX_INGREDIENT; i++)
 			{
-				PicMaterialIndex[(int) i] = 0;
-				LblMaterialName[(int) i] = "";
-				LblMaterialAmount[(int) i] = "";
+				PicMaterialIndex[i] = 0;
+				LblMaterialName[i] = "";
+				LblMaterialAmount[i] = "";
 			}
 			
 			CraftTimerEnabled = false;
